backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=4.4,<5.0
Pillow>=11.0,<12.0
gunicorn>=22.0,<24.0
whitenoise>=6.8,<7.0
